Extract from_user_data to a new ffi module in mruby crate
This function is unsafe and scary since it fiddles with a raw c_void userdata
pointer. Make it look scary.
