+++ mruby/src/interpreter.rs
@@ -1,7 +1,6 @@
-use log::{debug, error, trace};
+use log::{debug, error};
 use std::cell::RefCell;
 use std::ffi::c_void;
-use std::mem;
 use std::rc::Rc;
 
 use crate::eval::MrbEval;
@@ -54,33 +53,6 @@ impl Interpreter {
         interp.full_gc();
         Ok(interp)
     }
-
-    // TODO: Add a benchmark to make sure this function does not leak memory.
-    pub unsafe fn from_user_data(mrb: *mut sys::mrb_state) -> Result<Mrb, MrbError> {
-        if mrb.is_null() {
-            error!("Attempted to extract Mrb from null mrb_state");
-            return Err(MrbError::Uninitialized);
-        }
-        let ptr = (*mrb).ud;
-        if ptr.is_null() {
-            error!("Attempted to extract Mrb from null mrb_state->ud pointer");
-            return Err(MrbError::Uninitialized);
-        }
-        // Extract the smart pointer that wraps the API from the user data on
-        // the mrb interpreter. The `mrb_state` should retain ownership of its
-        // copy of the smart pointer.
-        let ud = Rc::from_raw(ptr as *const RefCell<State>);
-        // Clone the API smart pointer and increase its ref count to return a
-        // reference to the caller.
-        let api = Rc::clone(&ud);
-        // Forget the transmuted API extracted from the user data to make sure
-        // the `mrb_state` maintains ownership and the smart pointer does not
-        // get deallocated before `mrb_close` is called.
-        mem::forget(ud);
-        // At this point, `Rc::strong_count` will be increased by 1.
-        trace!("Extracted Mrb from user data pointer on {}", mrb.debug());
-        Ok(api)
-    }
 }
 
 #[cfg(test)]
@@ -88,38 +60,6 @@ mod tests {
     use crate::eval::MrbEval;
     use crate::interpreter::Interpreter;
     use crate::sys;
-    use crate::MrbError;
-
-    #[test]
-    fn from_user_data_null_pointer() {
-        unsafe {
-            let err = Interpreter::from_user_data(std::ptr::null_mut());
-            assert_eq!(err.err(), Some(MrbError::Uninitialized));
-        }
-    }
-
-    #[test]
-    fn from_user_data_null_user_data() {
-        unsafe {
-            let interp = Interpreter::create().expect("mrb init");
-            let api = interp.borrow();
-            let mrb = api.mrb;
-            // fake null user data
-            (*mrb).ud = std::ptr::null_mut();
-            let err = Interpreter::from_user_data(mrb);
-            assert_eq!(err.err(), Some(MrbError::Uninitialized));
-        }
-    }
-
-    #[test]
-    fn from_user_data() {
-        unsafe {
-            let interp = Interpreter::create().expect("mrb init");
-            let mrb = interp.borrow().mrb;
-            let res = Interpreter::from_user_data(mrb);
-            assert!(res.is_ok());
-        }
-    }
 
     #[test]
     fn open_close() {
